requirements.txt
torch
fairseq
gradio
transformers
Pillow
g2p-en
